refactor(time-tracker): migrate all types to feature scoped index
