+++ src/features/time-tracker/components/TimeEntryForm/ActivityTypeSelect.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import type { ActivityType } from "@/types/time-tracker.types";
-import { ACTIVITY_TYPE_OPTIONS } from "@/types/time-tracker.types";
+import type { ActivityType } from "@/features/time-tracker/types";
+import { ACTIVITY_TYPE_OPTIONS } from "@/features/time-tracker/types";
 
 interface ActivityTypeSelectProps {
   disabled?: boolean;
