+++ src/features/time-tracker/components/TimeRecordsList/SearchFilters.tsx
@@ -1,7 +1,7 @@
 import { Filter, Search } from "lucide-react";
 import React from "react";
 
-import { ACTIVITY_TYPE_OPTIONS, type ActivityType } from "@/types/time-tracker.types";
+import { ACTIVITY_TYPE_OPTIONS, type ActivityType } from "@/features/time-tracker/types";
 
 interface SearchFiltersProps {
   filterType: "" | ActivityType;
