+++ src/features/time-tracker/components/WeeklyView/DaySection.tsx
@@ -4,8 +4,8 @@ import { Calendar, Clock } from "lucide-react";
 import dynamic from "next/dynamic";
 import React, { useMemo } from "react";
 
-import type { TimeRecord } from "@/types/time-tracker.types";
-import { ActivityType } from "@/types/time-tracker.types";
+import type { TimeRecord } from "@/features/time-tracker/types";
+import { ActivityType } from "@/features/time-tracker/types";
 
 import { formatMinutesToHours, getActivityTypeColor } from "../../utils/formatting";
