+++ src/features/time-tracker/components/WeeklyView/WeeklyView.tsx
@@ -3,7 +3,7 @@
 import { Calendar } from "lucide-react";
 import React, { useMemo } from "react";
 
-import type { TimeRecord } from "@/types/time-tracker.types";
+import type { TimeRecord } from "@/features/time-tracker/types";
 
 import { useUserSettings } from "../../hooks";
 import { getWeeklyCategoryTotals } from "../../utils/statisticsCalculation";
